# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ZipCodes",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://backbonesystems.herokuapp.com"
        }
    ],
    "paths": {
...
        "/api/zip-codes/{zip_code}": {
            "get": {
                "tags": [
                    "ZipCodes"
                ],
                "summary": "Mostrar un ZipCodes",
                "operationId": "0efe6bfd1dedd5f3770ee6bdd3af8873",
                "parameters": [
                    {
                        "parameter": "ZipCode_in_query",
                        "name": "zip_code",
                        "in": "query",
                        "description": "Especifique el ZipCode",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "d_codigo": {
                                            "type": "string",
                                            "example": "1000"
                                        },
                                        "d_asenta": {
                                            "type": "string",
                                            "example": "San Ángel"
                                        },
                                        "d_tipo_asenta": {
                                            "type": "number",
                                            "example": "Colonia"
                                        },
                                        "d_mnpio": {
                                            "type": "string",
                                            "example": "Álvaro Obregón"
                                        },
                                        "d_estado": {
                                            "type": "string",
                                            "example": "Ciudad de México"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ZipCodes"
                ],
                "summary": "Eliminar ZipCodes",
                "operationId": "5692e4d82d633d366ff18ec52c0b4f95",
                "parameters": [
                    {
                        "parameter": "ZipCode_in_query",
                        "name": "zip_code",
                        "in": "query",
                        "description": "Especifique el ZipCode",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "d_codigo": {
                                            "type": "string",
                                            "example": "1000"
                                        },
                                        "d_asenta": {
                                            "type": "string",
                                            "example": "San Ángel"
                                        },
                                        "d_tipo_asenta": {
                                            "type": "number",
                                            "example": "Colonia"
                                        },
                                        "d_mnpio": {
                                            "type": "string",
                                            "example": "Álvaro Obregón"
                                        },
                                        "d_estado": {
                                            "type": "string",
                                            "example": "Ciudad de México"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}